feat: add enforce-check-sat flag

+++ dlinear/util/ArgParser.h
@@ -79,7 +79,7 @@ class ArgParser {
   friend std::ostream &operator<<(std::ostream &os, const ArgParser &parser);
 
   argparse::ArgumentParser parser_;  ///< The parser object.
-  size_t verbosity_;                 ///< Verbosity level of the program
+  int verbosity_;                    ///< Verbosity level of the program
   const std::string qsoptex_hash_;   ///< The hash of the QSoptEx library. Used in the prompt
   const std::string soplex_hash_;    ///< The hash of the Soplex library. Used in the prompt
 };
